CrossTab/requirements.txt
absl-py
aiohttp
aiosignal
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
anyio
appdirs
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-timeout
attrs
Babel
backcall
backports.functools-lru-cache
batchgenerators==0.25.1
beautifulsoup4
bgen-reader
bleach
blinker
bokeh
Brotli
brotlipy
cached-property
cachetools==4.2.4
carbontracker==1.1.6
cbgen
certifi==2022.12.7
cffi
charset-normalizer
click==8.1.8
cloudpickle
codecarbon
colorama
comm
configparser
contourpy
cryptography
cycler
cytoolz
dash
dash-bootstrap-components
dash-core-components
dash-html-components
dash-table
dask
debugpy
decorator
defusedxml
Deprecated
dicom2nifti==2.5.1
distributed
dnspython
docker-pycreds==0.4.0
einops==0.7.0
entrypoints
exceptiongroup
executing
fastjsonschema
filelock==3.17.0
fire
Flask==1.1.2
Flask-Compress
flit_core
fonttools
frozenlist
fsspec==2025.2.0
future
fuzzywuzzy
geocoder==1.38.1
gitdb
GitPython
google-auth==1.35.0
google-auth-oauthlib
grpcio
h5py
HeapDict==1.0.1
huggingface-hub==0.29.0
humanize
hydra-core==1.2.0
icd10-cm==0.0.5
idna
imagecodecs
imageio
importlib-metadata
importlib-resources
iniconfig
ipykernel
ipython
ipython-genutils==0.2.0
itsdangerous
jedi
Jinja2
joblib
json5
jsonschema
jupyter-server
jupyter_client
jupyter_core
jupyterlab
jupyterlab-pygments
jupyterlab_server
kiwisolver
Levenshtein
liac-arff==2.5.0
lightly==1.2.22
lightly-utils==0.0.2
lightning-bolts==0.5.0
lime
linecache2==1.0.0
llvmlite==0.39.1
locket
lz4
Markdown
markdown-it-py==3.0.0
MarkupSafe
matplotlib
matplotlib-inline
matplotlib-venn==0.11.7
mdurl==0.1.2
MedPy==0.5.2
minio==7.1.9
mistune
monai
msgpack
multidict
munkres==1.1.4
mysql-connector-python==8.0.31
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nibabel
notebook_shim
numba
numpy==1.24
oauthlib
omegaconf==2.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openml==0.12.2
packaging
pandas==1.5.2
pandocfilters
parso
partd
pathtools==0.1.2
patsy
pexpect
pickleshare
Pillow
pkgutil_resolve_name
platformdirs
plotly
pluggy
pooch
prometheus-client
promise
prompt-toolkit
protobuf==3.20.1
psutil
ptyprocess
pure-eval
py-cpuinfo
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser
pyDeprecate
pydicom
Pygments
PyJWT
pynndescent
pynvml
pyOpenSSL
pyparsing
pyrsistent
PySocks
pytest==7.2.0
python-dateutil
python-gdcm==3.0.24.1
python-Levenshtein
pytorch-lightning
pytz
pyu2f
PyWavelets
PyYAML
pyzmq
qudida==0.0.4
rapidfuzz
ratelim==0.1.6
regex==2024.11.6
requests
requests-oauthlib
rich==13.9.4
rsa
safetensors==0.5.2
scikit-image
scikit-learn
scipy==1.13.1
seaborn
Send2Trash
sentry-sdk
setproctitle
shellingham==1.5.4
shortuuid
simple-icd-10==2.0.1
SimpleITK==2.2.1
six
smmap
sniffio
sortedcontainers
soupsieve
stack-data
statsmodels
tabpfn==0.1.8
tblib
tenacity
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
termcolor
terminado
texttable
threadpoolctl
tifffile
timm==0.9.12
tinycss2
tokenizers==0.21.0
tomli
toolz
torch==1.11.0
torch-tb-profiler==0.4.0
torchaudio==0.11.0
torchio
torchmetrics
torchvision==0.12.0
tornado
tqdm
traceback2==1.4.0
traitlets
transformers==4.49.0
typer==0.15.2
typing_extensions
umap-learn
unicodedata2
unittest2==1.1.0
urllib3
wandb
wcwidth
webencodings==0.5.1
websocket-client
Werkzeug
wrapt
xarray
xgboost==1.7.4
xmltodict==0.13.0
yarl
yaspin
zict
zipp